refactoring: extract configuration common code into single place
