+++ Source/EasyGelf.NLog/GelfUdpTarget.cs
@@ -1,5 +1,4 @@
-﻿using System.Linq;
-using System.Net;
+﻿using System.Net;
 using EasyGelf.Core;
 using EasyGelf.Core.Encoders;
 using EasyGelf.Core.Transports;
@@ -27,13 +26,10 @@ namespace EasyGelf.NLog
         protected override ITransport InitializeTransport(IEasyGelfLogger logger)
         {
             var encoder = new CompositeEncoder(new GZipEncoder(), new ChunkingEncoder(new MessageBasedIdGenerator(), MessageSize.UdpMessageSize()));
-            var removeIpAddress = Dns.GetHostAddresses(RemoteAddress)
-                .Shuffle()
-                .DefaultIfEmpty(IPAddress.Loopback)
-                .First();
             var configuration = new UdpTransportConfiguration
                 {
-                    Host = new IPEndPoint(removeIpAddress, RemotePort)
+                    RemoteAddress = RemoteAddress,
+                    RemotePort = RemotePort
                 };
             return new UdpTransport(configuration, encoder, new GelfMessageSerializer());
         }
